+++ Magic/CHANGELOG.md
@@ -36,6 +36,7 @@
  - Add `not_brush` parameter to `CheckBlock`, use-case is only playing effects at a location where a block was modified (not the same as the brush)
  - The `ColorItem` action is now *mostly* undoable, though probably only suited for mob use (since players can move items around)
  - Add `swappable_slots` wand property, a list of slot types that can have their slotted items swapped out
+ - Add `magic_blocks` list to arenas, arena will turn the magic block on/off when the arena starts/stops
 
 ## Changes